Add AOT versions of NativeExecutor and ProgramCache.
