+++ src/utils.rs
@@ -28,6 +28,11 @@ use std::{
 };
 use thiserror::Error;
 
+#[cfg(target_os = "macos")]
+pub const SHARED_LIBRARY_EXT: &str = "dylib";
+#[cfg(target_os = "linux")]
+pub const SHARED_LIBRARY_EXT: &str = "so";
+
 /// Generate a function name.
 ///
 /// If the program includes function identifiers, return those. Otherwise return `f` followed by the
